+++ src/test/system/quick/test_small_db_install.py
@@ -22,7 +22,7 @@ import os
 import os.path
 from pathlib import Path
 from downloader.config import ConfigReader
-from test.objects import default_env
+from test.objects import debug_env
 from test.fakes import NoLogger
 import subprocess
 
@@ -34,7 +34,7 @@ class TestSmallDbInstall(unittest.TestCase):
         self.assertRunOk("test/system/fixtures/small_db_install/small_db.ini")
 
     def assertRunOk(self, ini_path):
-        config = ConfigReader(NoLogger(), default_env()).read_config(ini_path)
+        config = ConfigReader(NoLogger(), debug_env()).read_config(ini_path)
         shutil.rmtree(config['base_path'], ignore_errors=True)
         shutil.rmtree(config['base_system_path'], ignore_errors=True)
         mister_path = Path('%s/MiSTer' % config['base_system_path'])
@@ -46,6 +46,7 @@ class TestSmallDbInstall(unittest.TestCase):
         subprocess.run(['chmod', '+x', tool], shell=False, stderr=subprocess.STDOUT)
         test_env = os.environ.copy()
         test_env['CURL_SSL'] = ''
+        test_env['DEBUG'] = 'true'
         result = subprocess.run([tool], stderr=subprocess.STDOUT, env=test_env)
         self.assertEqual(result.returncode, 0)
         self.assertTrue(os.path.isfile("%s/Scripts/.config/downloader/downloader.json.zip" % config['base_system_path']))